// frontend/src/app/core/model/employee.ts
// Copyright 2016 Google Inc. All Rights Reserved.
// Use of this source code is governed by an MIT-style license that can be found in the LICENSE file at https://angular.io/license
                    

export class Employee {
  id:string;
  name: string;
  designation: string;
  gender: string;
  email: string;
  department:string;
}

 


// Copyright 2016 Google Inc. All Rights Reserved.
// Use of this source code is governed by an MIT-style license that can be found in the LICENSE file at https://angular.io/license
            